+++ Realm/RLMArray.mm
@@ -146,7 +146,7 @@ static void RLMValidateMatchingObjectType(RLMArray *array, RLMObject *object) {
 
 - (void)deleteObjectsFromRealm {
     for (RLMObject *obj in _backingArray) {
-        RLMDeleteObjectFromRealm(obj);
+        RLMDeleteObjectFromRealm(obj, _realm);
     }
 }