+++ CHANGELOG.md
@@ -1,4 +1,43 @@
-x.x.x Release notes (yyyy-MM-dd)
+0.88.0 Release notes (YYYY-MM-DD)
+=============================================================
+
+### API breaking changes
+
+* None.
+
+### Enhancements
+
+* Add `-[RLMRealm writeCopyToPath:]` to write a compacted copy of the Realm
+  another file.
+* Add support for case insensitive, BEGINSWITH, ENDSWITH and CONTAINS string
+  queries on array properties.
+* Make fast enumeration of `RLMArray` and `RLMResults` ~30% faster and
+  `objectAtIndex:` ~55% faster.
+* Added a lldb visualizer script for displaying the contents of persisted
+  RLMObjects when debugging.
+
+### Bugfixes
+
+* Fix for crash when running many simultaneous write transactions on background threads.
+* Don't run the query twice when `firstObject` or `lastObject` are called on an
+  `RLMResults` which has not had its results accessed already.
+
+0.87.4 Release notes (2014-11-07)
+=============================================================
+
+### API breaking changes
+
+* None.
+
+### Enhancements
+
+* None.
+
+### Bugfixes
+
+* Fix browser location in release zip.
+
+0.87.3 Release notes (2014-11-06)
 =============================================================
 
 ### API breaking changes
@@ -8,7 +47,7 @@ x.x.x Release notes (yyyy-MM-dd)
 ### Enhancements
 
 * Added method `-linkingObjectsOfClass:forProperty:` to RLMObject to expose inverse
-  relashinships/backlinks.
+  relationships/backlinks.
 
 ### Bugfixes
 
@@ -16,7 +55,7 @@ x.x.x Release notes (yyyy-MM-dd)
   in a database created using an older versions (0.86.3 and earlier).
 * Throw an exception when passing an array containing a
   non-RLMObject to -[RLMRealm addObjects:].
-
+* Fix for crash when deleting an object from multiple threads.
 
 0.87.0 Release notes (2014-10-21)
 =============================================================